+++ Core/Entities/Block/Auditor.cs
@@ -9,7 +9,7 @@ namespace Core.Entities.Block;
 public class Auditor : BaseEntity
 {
     public string Name { get; set; }
-    public long ActionDes { get; set; }
+    public int ActionDes { get; set; }
     public DateTime CreationDate { get; set; }
     public DateTime ModificationDate { get; set; }
